smoke-tests/requirements.txt
pytest>=7.4,<9
pytest-ordering>=0.6
requests>=2.31,<3
azure-identity>=1.15,<2
python-dotenv>=1.0,<2
